Move test files out of src dir
