+++ packages/parser-buffer/tsconfig.json
@@ -5,5 +5,6 @@
     "outDir": "lib",
     "rootDir": "src"
   },
+  "include": ["src"],
   "references": [{ "path": "../parser" }]
 }
