+++ packages/parser/tsconfig.json
@@ -4,5 +4,6 @@
     "composite": true,
     "outDir": "lib",
     "rootDir": "src"
-  }
+  },
+  "include": ["src"]
 }
